Fixed double/float Cython issue. Starting new types for Saddles and Modes

+++ video_proj/indexing.pxd
@@ -16,6 +16,6 @@ cdef np.ndarray[idx_type, ndim=1] flatten_idc(
     np.ndarray[idx_type, ndim=1] dims
     )
 
-cdef inline idx_type clamp(float n, idx_type mn, idx_type mx)
+cdef inline idx_type clamp(double n, idx_type mn, idx_type mx)
 
 cdef inline int oob(idx_type i, idx_type N)
